Move DevTools out of flutter
