+++ src/commands/debug.ts
@@ -3,9 +3,9 @@ import * as path from "path";
 import * as vs from "vscode";
 import { Analytics } from "../analytics";
 import { CoverageData, PromiseCompleter } from "../debug/utils";
-import { FlutterDevTools } from "../flutter/dev_tools";
 import { FlutterServiceExtension, FlutterServiceExtensionArgs, FlutterVmServiceExtensions, timeDilationNormal, timeDilationSlow } from "../flutter/vm_service_extensions";
 import { PubGlobal } from "../pub/global";
+import { FlutterDevTools } from "../sdk/dev_tools";
 import { fsPath, getDartWorkspaceFolders, openInBrowser, Sdks } from "../utils";
 import { DartDebugSessionInformation } from "../utils/debug";
 import { handleDebugLogEvent } from "../utils/log";
